Backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
bcrypt
python-dotenv
requests
pillow
httpx
transformers
torch
scikit-learn
joblib
numpy

